# Minor Project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/admin/details": {
            "get": {
                "tags": [
                    "admin details"
                ],
                "summary": "admin details",
                "operationId": "details",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/company/forms_submit": {
            "post": {
                "tags": [
                    "submit jobform"
                ],
                "summary": "submit new form to add a new job post",
                "operationId": "submit_job_form",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "company_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "skills": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/job_posted": {
            "get": {
                "tags": [
                    "jobs_posted"
                ],
                "summary": "get all skills",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "get_all_post",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/company/update_job_forms/{job_id}": {
            "get": {
                "tags": [
                    "update job form display"
                ],
                "summary": "update job form display",
                "operationId": "update_job_form_display",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "job id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/company/update_job_form/{job_id}": {
            "post": {
                "tags": [
                    "update jobform"
                ],
                "summary": "update job form to add edit job post",
                "operationId": "update_job_form",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job form id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "company_id": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "skills": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/company/delete_job_form/{job_id}": {
            "get": {
                "tags": [
                    "delete job form"
                ],
                "summary": "delete job form posted by a given company",
                "operationId": "delete_job_form",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "job id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/company/job_list": {
            "post": {
                "tags": [
                    "get job posted by given company"
                ],
                "summary": "get job posted by given company",
                "operationId": "job_list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/expired_job_list": {
            "post": {
                "tags": [
                    "get expired job posted by given company"
                ],
                "summary": "get expired job posted by given company",
                "operationId": "get_exp_job_list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/expired_job_list/{job_id}": {
            "post": {
                "tags": [
                    "get all user form from a expired job"
                ],
                "summary": "get all user form for a given expired job post",
                "operationId": "get_applied_user_form_exp_job",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/expired_job_list/pdf/{job_id}/{id}": {
            "post": {
                "tags": [
                    "get user form pdf from exp job"
                ],
                "summary": "get user form  pdf for a given job post",
                "operationId": "user_exp_form_pdf",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "exp job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "exp user form id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/company/expired_job_user_detail/{job_id}/{id}": {
            "post": {
                "tags": [
                    "get user details"
                ],
                "summary": "get user form  details who has applied for a given job post tht has been expired",
                "operationId": "user_details_exp",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "job post id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "default": "available"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}